Adding CSV as repository

+++ src/test/java/org/example/ProductDefinitionTest.java
@@ -7,6 +7,7 @@ import org.example.product.ProductDefinition;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import static org.assertj.core.api.Assertions.*;
@@ -19,7 +20,7 @@ class ProductDefinitionTest {
     @SuppressWarnings("DataFlowIssue")
     @Test
     void checkingIf_whenAddedNewProductWithNullProductName_thenMethodThrowsException() {
-        assertThatThrownBy(() -> new ProductDefinition(null, new Price(Money.of(2.50, PLN)), creationDate))
+        assertThatThrownBy(() -> new ProductDefinition(null, new Price(Money.of(BigDecimal.valueOf(2.50), PLN)), creationDate))
                 .hasMessage("You cannot add or remove product with null name");
     }
